fix msvc compilation warning

+++ hardware/Tellstick.cpp
@@ -66,13 +66,13 @@ void CTellstick::sensorEvent(int deviceId, const char *protocol, const char *mod
     switch (dataType)
     {
     case TELLSTICK_TEMPERATURE:
-        SendTempSensor(deviceId, 255, atof(value), "Temp");
+        SendTempSensor(deviceId, 255, (float) atof(value), "Temp");
         break;
     case TELLSTICK_HUMIDITY:
         SendHumiditySensor(deviceId, 255, atoi(value), "Humidity");
         break;
     case TELLSTICK_RAINRATE:
-        SendRainSensor(deviceId, 255, atof(value), "Rain");
+        SendRainSensor(deviceId, 255, (float) atof(value), "Rain");
         break;
     case TELLSTICK_RAINTOTAL:
     case TELLSTICK_WINDDIRECTION:
